feat: decrypt all layers when message is echoed back to client
This feature decrypts all layers in order when the decrypted message is echoed back to the client.
As currently only the first relay is decrypting and being the only hop in the routing, for demo
purposes the message after the first hop is echoed back to the client and being decrypted by
the client.